Consolidated models into modules and updated plotting.
