+++ data/.gitignore
@@ -1,0 +1,3 @@
+*.csv
+*.gif
+
