+++ data/README.md
@@ -1,0 +1,1 @@
+Just a directory to hold model and plotting outputs.
